requirements.txt
pillow
transformers>=4.41.0
torch
pytesseract
openai>=1.0.0
